BalanceChangesReplacer: remove zlib1.dll dependency
The zlib1.dll which was previously included was a 64-bit binary.
I didn't notice this mistake until today, when I decided to do some research...
The alternatives were: include the correct 32-bit binary, or statically link zlib.
I went for the third option: to use a compression function already available in
the game's code.